Add gray opt to kcw and switch to Y800 codec
Added a control option to the KeyClipWriter class to allow for saving
grayscale images. The default is still True so it should not effect any
existing programs.

Also am trying to switch to the uncompressed Y800 uint8 codec, which
seems to have much faster write times (likely owing to no transcoding)
at the cost of larger files. False positives seems low enough for this
to be viable, but could always transcode during daylight hours if space
really becomes an issue.